Moved handlers to handler space
Moved the handlers to the handler space for organisation
